--- a/packages/backend/discovery/apex/ethereum/diffHistory.md
+++ b/packages/backend/discovery/apex/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0x7a6827f31d178614aac9c3b4e009541e8656e6cf
+Generated with discovered.json: 0x9efaa4125824a49cd3c3eba891f97f65c0d9b99f
 
 # Diff at Mon, 20 Jan 2025 11:09:15 GMT: