--- a/packages/backend/discovery/astarzkevm/ethereum/diffHistory.md
+++ b/packages/backend/discovery/astarzkevm/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0x87fda66836b980612621d681b3af04aec35db16c
+Generated with discovered.json: 0xab3c0a75e7c9774b5a715df68694eec9ad18adf9
 
 # Diff at Mon, 20 Jan 2025 11:09:17 GMT: