--- a/packages/backend/discovery/brine/ethereum/diffHistory.md
+++ b/packages/backend/discovery/brine/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0xb3f7d8aa355f411b1963d446c58004945b1a2d21
+Generated with discovered.json: 0xe676b810d2c6f4b8ff9601b69d8960bdd7c1718a
 
 # Diff at Mon, 20 Jan 2025 11:09:21 GMT: