--- a/packages/backend/discovery/canvasconnect/ethereum/diffHistory.md
+++ b/packages/backend/discovery/canvasconnect/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0x54c2eb77c8c01712cdf30477003c5eaa7978f1e6
+Generated with discovered.json: 0x941a7677e2a50eb0c2b6ac2cb520237bb006735f
 
 # Diff at Mon, 20 Jan 2025 11:09:21 GMT: