--- a/packages/backend/discovery/canvasconnect/ethereum/discovered.json
+++ b/packages/backend/discovery/canvasconnect/ethereum/discovered.json
@@ -3,6 +3,7 @@
   "chain": "ethereum",
   "blockNumber": 19825363,
   "configHash": "0x1186dd80cc3f9795816bb690b7faadee1864ab17eba5d25e1af47d50fa8f4def",
+  "sharedModules": ["shared-sharp-verifier"],
   "contracts": [
     {
       "name": "GpsFactRegistryAdapter",