--- a/packages/backend/discovery/cronoszkevm/ethereum/diffHistory.md
+++ b/packages/backend/discovery/cronoszkevm/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0xe0cd4967c23c2f9e9be33cb5f0e793a33fbd39f2
+Generated with discovered.json: 0x5e8e0d515eadb4b77c41c006266778320ea52b99
 
 # Diff at Thu, 23 Jan 2025 09:37:44 GMT: