--- a/packages/backend/discovery/deversifi/ethereum/diffHistory.md
+++ b/packages/backend/discovery/deversifi/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0x93f3ff8da78c4a5bab6d0d0e91584dd0216261e0
+Generated with discovered.json: 0xdd684683f170aba7a9fcfcc44fe2e42b02e9226b
 
 # Diff at Mon, 20 Jan 2025 11:09:25 GMT: