--- a/packages/backend/discovery/eigenda/ethereum/diffHistory.md
+++ b/packages/backend/discovery/eigenda/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0xd4f9f1f4c46984e287ab322a4c0b93a22d3fe85a
+Generated with discovered.json: 0xfdecefca0505b39cd2ee474098d469339feb25d1
 
 # Diff at Mon, 27 Jan 2025 08:44:35 GMT: