--- a/packages/backend/discovery/gpt/ethereum/diffHistory.md
+++ b/packages/backend/discovery/gpt/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0xfefa2eea1f1591cc7d48c4d83ed641816edeba5f
+Generated with discovered.json: 0x20bd0ffd0c9b9e7795b3bc9c9774291626f20f85
 
 # Diff at Mon, 20 Jan 2025 11:09:33 GMT: