--- a/packages/backend/discovery/grvt/ethereum/diffHistory.md
+++ b/packages/backend/discovery/grvt/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0x332f9b98d8c5fd10db3ff8890dca23f240f438cd
+Generated with discovered.json: 0x8389fd0febf49490dfb76c31b368eb6fbf1d8530
 
 # Diff at Mon, 20 Jan 2025 11:09:34 GMT: