--- a/packages/backend/discovery/immutablex/ethereum/diffHistory.md
+++ b/packages/backend/discovery/immutablex/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0x97e5279e959eb2e772c14be6234d447e46e863b9
+Generated with discovered.json: 0x578ad983ad1e920e972e8e745308c2ba122d4f40
 
 # Diff at Mon, 20 Jan 2025 11:09:36 GMT: