--- a/packages/backend/discovery/layer2financezk/ethereum/diffHistory.md
+++ b/packages/backend/discovery/layer2financezk/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0xba72cddecce686150b195cd33bc0a439426b42ef
+Generated with discovered.json: 0x51bc15d6dbdceaec2269107c28dc6213a72831b7
 
 # Diff at Mon, 20 Jan 2025 11:09:40 GMT: