--- a/packages/backend/discovery/myria/ethereum/diffHistory.md
+++ b/packages/backend/discovery/myria/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0xc9fd30041ffebfec1fb53663f78ef32beb874dfe
+Generated with discovered.json: 0x692a8273fb1be8fe1333860c99ae135d56a61444
 
 # Diff at Mon, 20 Jan 2025 11:09:47 GMT: