--- a/packages/backend/discovery/optimism/ethereum/diffHistory.md
+++ b/packages/backend/discovery/optimism/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0x87efb2585e13b00ac3aae3268faa047ddcfe14a3
+Generated with discovered.json: 0xfe4f950cc9481767f7f4d17684d9295c599d211a
 
 # Diff at Tue, 21 Jan 2025 11:19:15 GMT: