--- a/packages/backend/discovery/optimism/ethereum/discovered.json
+++ b/packages/backend/discovery/optimism/ethereum/discovered.json
@@ -318,7 +318,7 @@
       "values": {
         "$immutable": true,
         "safe": "0xc2819DC788505Aac350142A7A707BF9D03E3Bd03",
-        "timeSinceLastL2beatInteraction": "9d 16h",
+        "timeSinceLastL2beatInteraction": "9d 17h",
         "version": "1.0.0"
       },
       "derivedName": "LivenessGuard"