--- a/packages/backend/discovery/paradex/ethereum/diffHistory.md
+++ b/packages/backend/discovery/paradex/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0x2d8ed074a98b6921ca5b8c5aa2a2a873a6fff42d
+Generated with discovered.json: 0x7543e3797229957078655ecd9be43d97603a96ef
 
 # Diff at Mon, 20 Jan 2025 11:09:53 GMT: