--- a/packages/backend/discovery/paradex/ethereum/discovered.json
+++ b/packages/backend/discovery/paradex/ethereum/discovered.json
@@ -3,6 +3,7 @@
   "chain": "ethereum",
   "blockNumber": 21629810,
   "configHash": "0xaa0aa74ef8fd5b854c9bdc4faf84f21c2f051ce5ee9549d332769fbbf7617fda",
+  "sharedModules": ["shared-sharp-verifier"],
   "contracts": [
     {
       "name": "ParadexImplementationGovernorMultisig",