--- a/packages/backend/discovery/polygonzkevm/ethereum/diffHistory.md
+++ b/packages/backend/discovery/polygonzkevm/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0xaf1789d5aa68cdb95fdbced8505a71d082d6b9d3
+Generated with discovered.json: 0xf75a5eb2457d27bc6bf4f0c93ba7130daee9f7d4
 
 # Diff at Mon, 20 Jan 2025 11:09:55 GMT: