--- a/packages/backend/discovery/reddioex/ethereum/diffHistory.md
+++ b/packages/backend/discovery/reddioex/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0x1f921f53080ce1afbe2e86c15ee4a8183a83d969
+Generated with discovered.json: 0x3eaee540f9209ace5924e0e769ae18059d33d122
 
 # Diff at Mon, 20 Jan 2025 11:09:58 GMT: