--- a/packages/backend/discovery/reddioex/ethereum/discovered.json
+++ b/packages/backend/discovery/reddioex/ethereum/discovered.json
@@ -3,6 +3,7 @@
   "chain": "ethereum",
   "blockNumber": 19825385,
   "configHash": "0x52fd208c42da4bb6c8289925e32c4fc4d9a65cbf653f16bf3398570f09b7c047",
+  "sharedModules": ["shared-sharp-verifier"],
   "contracts": [
     {
       "name": "Committee",