--- a/packages/backend/discovery/silicon/ethereum/diffHistory.md
+++ b/packages/backend/discovery/silicon/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0x6d0e061c9405d036a0af0ec1c2123438eda83e85
+Generated with discovered.json: 0xcd5e6c347866f9b8b557e681f526ac57d516b5e9
 
 # Diff at Mon, 20 Jan 2025 11:10:06 GMT: