--- a/packages/backend/discovery/sophon/ethereum/diffHistory.md
+++ b/packages/backend/discovery/sophon/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0x58a5ad3c6d5983fe0f2d1a9f12b354e06c7125e7
+Generated with discovered.json: 0x5eb22ac8d74bc92d3dbf90a46549a36c36d4706c
 
 # Diff at Wed, 22 Jan 2025 12:22:26 GMT: