--- a/packages/backend/discovery/sorare/ethereum/diffHistory.md
+++ b/packages/backend/discovery/sorare/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0x47d4a4e8c7606ccd508752730dfa097f43f91dfd
+Generated with discovered.json: 0x27954a37dd949dd2df85ccc240885179573f5185
 
 # Diff at Mon, 20 Jan 2025 11:10:10 GMT: