--- a/packages/backend/discovery/starknet/ethereum/diffHistory.md
+++ b/packages/backend/discovery/starknet/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0xe1a151087c3ce613b22d4e8c7f7f5000aa0fb778
+Generated with discovered.json: 0xb9503031051fa02f99e862e89799c59f17f066ff
 
 # Diff at Mon, 20 Jan 2025 11:10:12 GMT: