--- a/packages/backend/discovery/treasure/ethereum/diffHistory.md
+++ b/packages/backend/discovery/treasure/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0xc04b8f4486721901d4ec338ed71da8a2fa993519
+Generated with discovered.json: 0x9ccc58c9872fd26aaf70224eff95e611fe660852
 
 # Diff at Thu, 09 Jan 2025 06:58:54 GMT: