--- a/packages/backend/discovery/wirex/ethereum/diffHistory.md
+++ b/packages/backend/discovery/wirex/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0x5d14ed154ff97a54cc3da8cbcbef1ae7b498fd3c
+Generated with discovered.json: 0xfb73a1a9b97e5619b3a87225bc6aa537141c27b2
 
 # Diff at Mon, 20 Jan 2025 11:10:20 GMT: