--- a/packages/backend/discovery/witness/ethereum/diffHistory.md
+++ b/packages/backend/discovery/witness/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0xaecb72b56f4253812a7f9aa24d19de3d1e66a9e2
+Generated with discovered.json: 0xf62412de983d39a423fea62956f99ee7593c1cb0
 
 # Diff at Mon, 20 Jan 2025 11:10:20 GMT: