--- a/packages/backend/discovery/xlayer/ethereum/diffHistory.md
+++ b/packages/backend/discovery/xlayer/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0x750011224acc0426108e062745e9f260a8b59a24
+Generated with discovered.json: 0x840195c017aacfa4b6a66094bc3d25659296fc17
 
 # Diff at Mon, 20 Jan 2025 12:06:34 GMT: