--- a/packages/backend/discovery/zeronetwork/ethereum/diffHistory.md
+++ b/packages/backend/discovery/zeronetwork/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0x056eb47d3a636fbda5fffd20b0031ffa242bc03c
+Generated with discovered.json: 0x096276bf47fe0662e7b13e891c951db44428de7c
 
 # Diff at Fri, 17 Jan 2025 08:39:46 GMT: