--- a/packages/backend/discovery/zksync2/ethereum/diffHistory.md
+++ b/packages/backend/discovery/zksync2/ethereum/diffHistory.md
@@ -1,4 +1,4 @@
-Generated with discovered.json: 0xbe786efbf76e30a304dd44b399b0d5fc4780a788
+Generated with discovered.json: 0x36b8c6d9ead866cfaf46cdf38330c2c22943315b
 
 # Diff at Wed, 15 Jan 2025 13:55:12 GMT: