--- a/packages/discovery-types/src/Discovery.ts
+++ b/packages/discovery-types/src/Discovery.ts
@@ -29,6 +29,7 @@ export interface DiscoveryOutput {
   eoas: EoaParameters[]
   abis: Record<string, string[]>
   configHash: Hash256
+  sharedModules?: string[]
   usedTemplates: Record<string, Hash256>
 }