Removes the need to read the config in ProjectDiscovery, saves time but not running Zod validation.
Tested with `hyperfine --warmup 5 "echo echo 'require("@l2beat/config")' | node"`

Before:

```
Benchmark 1: Config read
  Time (mean ± σ):      1.450 s ±  0.029 s    [User: 1.609 s, System: 0.205 s]
  Range (min … max):    1.417 s …  1.498 s    10 runs
```

After
```
Benchmark 1: Just discovery
  Time (mean ± σ):      1.185 s ±  0.028 s    [User: 1.261 s, System: 0.158 s]
  Range (min … max):    1.157 s …  1.237 s    10 runs
```